Avoid building unsupported darwin types on darwin
There are a few types and functions that modernc.org/libfreetype does
not support on darwin. They are now only used with linux builds.
