+++ glyph_retrieval_linux.go
@@ -1,0 +1,7 @@
+//go:build linux
+
+package freetype
+
+func (rec GlyphSlotRec) SVGDocument() *SVGDocumentRec {
+	return fromUintptr[SVGDocumentRec](rec.Other)
+}
