+++ opentype_svg_fonts.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package freetype
 
 import (
