+++ opentype_svg_fonts_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package freetype
 
 import (
